refine api docs and export split
Split the user export endpoint into its own module and tightened the public contract surface of the refactored API files.

- move auth export into a dedicated module to shrink account-related import chains
- add pydocs and tracing to the split-out public endpoints
- keep private helpers private while exposing documented traced wrappers for public contracts
- preserve compatibility imports for existing call sites and tests
