+++ api/BUILD.bazel
@@ -21,6 +21,7 @@ py_library(
         "analysis_views.py",
         "apps.py",
         "auth_account_views.py",
+        "auth_export_views.py",
         "dev_bootstrap.py",
         "auth_google_views.py",
         "auth_invite_views.py",
@@ -114,6 +115,7 @@ filegroup(
         "analysis_views.py",
         "apps.py",
         "auth_account_views.py",
+        "auth_export_views.py",
         "dev_bootstrap.py",
         "auth_google_views.py",
         "auth_invite_views.py",
