+++ api/cloudinary_admin_views.py
@@ -82,6 +82,7 @@ from backend.otel import traced
 @permission_classes([IsAdminUser])
 @traced
 def admin_cloudinary_cleanup(request: Request) -> Response:
+    """List or delete Cloudinary assets for admin cleanup."""
     from .cloudinary_cleanup import (
         delete_cloudinary_assets,
         list_cloudinary_assets,
@@ -175,6 +176,7 @@ def admin_cloudinary_cleanup(request: Request) -> Response:
 def admin_cloudinary_cleanup_archive(
     request: Request,
 ) -> StreamingHttpResponse | Response:
+    """Download a ZIP archive of Cloudinary assets for admin cleanup."""
     from .cloudinary_cleanup import (
         list_cloudinary_assets,
         stream_cloudinary_cleanup_archive,
